+++ py/desitarget/streams/io.py
@@ -389,7 +389,8 @@ def read_data_per_stream(swdir, rapol, decpol, mind, maxd, stream_name,
 
 
 def write_targets(dirname, targs, header, nside=None, pixint=None,
-                  subpriority=True, nsidecol=None):
+                  subpriority=True, nsidecol=None, addstandards=False,
+                  backupdir=None):
     """Write stream and dwarf targets to a FITS file.
 
     Parameters
@@ -417,6 +418,13 @@ def write_targets(dirname, targs, header, nside=None, pixint=None,
     nsidecol : :class:`int`, optional, defaults to `None`
         If passed, add a column to the targets array popluated
         with HEALPixels in the nested scheme at resolution `nsidecol`.
+    addstandards : :class:`bool`, optional, defaults to ``False``
+        If ``True`` then add standard stars from the BACKUP program by
+        reading them from a directory of BACKUP files.
+    backupdir : :class:`str`, optional, defaults to `None`
+        Use this as the directory that hosts the BACKUP target files if
+        `addstandards` is ``True``. If `None` is passed, defaults to:
+        $TARG_DIR/gaiadr2/2.2.0/targets/main/resolve/backup
 
     Returns
     -------
@@ -430,8 +438,8 @@ def write_targets(dirname, targs, header, nside=None, pixint=None,
     Notes
     -----
     - Must contain at least the columns:
-        PHOT_G_MEAN_MAG, PHOT_BP_MEAN_MAG, PHOT_RP_MEAN_MAG and
-        FIBERTOTFLUX_G, FIBERTOTFLUX_R, FIBERTOTFLUX_Z, RELEASE
+        GAIA_PHOT_G_MEAN_MAG, GAIA_PHOT_BP_MEAN_MAG, GAIA_PHOT_RP_MEAN_MAG
+        and FIBERTOTFLUX_G, FIBERTOTFLUX_R, FIBERTOTFLUX_Z, RELEASE
     - Always OVERWRITES existing files!
     - Writes atomically. Any output files that died mid-write will be
       appended by ".tmp".
@@ -451,7 +459,9 @@ def write_targets(dirname, targs, header, nside=None, pixint=None,
 
     # ADM construct the output filename.
     drs = list(set(targs["RELEASE"]//1000))
-    if len(drs) == 1:
+    # ADM note that we special case RELEASE=7777, which corresponds to
+    # ADM the M31/M33 BRIGHT1B program.
+    if len(drs) == 1 and not np.all(targs["RELEASE"] == 7777):
         drint = drs[0]
         drstr = f"dr{drint}"
     else:
@@ -504,6 +514,12 @@ def write_targets(dirname, targs, header, nside=None, pixint=None,
         ii = targs["SUBPRIORITY"] == 0.0
         targs["SUBPRIORITY"][ii] = np.random.random(len(targs))[ii]
         header["SUBPSEED"] = subpseed
+
+    # ADM add standard stars for BACKUP program, if requested.
+    if addstandards:
+        from desitarget.streams.M31cuts import add_backup_standards
+        targs = add_backup_standards(targs, backupdir=backupdir,
+                                     nside=nside, pixnum=pixint)
 
     # ADM add the DESI dependencies.
     depend.add_dependencies(header)